Attempting to only select appointments that haven't already been made.

+++ src/schedule_appointment.php
@@ -9,7 +9,7 @@
     if(empty($_SESSION['user'])) {
         header("Location: ../index.php");
         die("Redirecting to index.php");
-    } else if (!empty($_POST['time'])) {
+    } else if (!empty($_POST['time']) && !empty($_POST['date']) && !empty($_POST['doctor_name'])) {
         $appointment = new ScheduleAppointment($_POST["doctor_name"], $_SESSION["user"]["first_name"]
                 . " " . $_SESSION["user"]["last_name"], $_SESSION["user"]["email"], $_POST["date"], $_POST["time"], $db);
         if (empty($appointment->error)) {
@@ -43,6 +43,8 @@
                 }
             }
         }
+    } else if (!empty($_POST['date']) && !empty($_POST['doctor_name'])) {
+        
     }
 ?>
 
@@ -64,11 +66,9 @@
     <script src="http://code.jquery.com/ui/1.11.4/jquery-ui.js"></script>
     <script>$(function() {$( "#datepicker" ).datepicker({minDate: "+1D", maxDate: "+6M", beforeShowDay: $.datepicker.noWeekends});});</script>
     <script>
-        $(function() {
-            $('#doctorNameUpdate').change(function() {
-                this.form.submit();
-            });
-        });
+        function doctorNameUpdated() {
+            document.getElementById("mainForm").submit();
+        }
         function dateUpdated() {
             document.getElementById("mainForm").submit();
         }
@@ -213,20 +213,86 @@
                     // E.g 19-3 => 19-27 for simplicity.
                     $endTime += 24;
                 }
+                // Determine appointments that have already been scheduled.
+                $query = '
+                        SELECT *
+                        FROM appointment
+                        WHERE
+                            doctor_name = :doctorName
+                            AND
+                            date = :date
+                        ';
+                $query_params = array(
+                    ':doctorName' => $_POST['doctor_name'],
+                    ':date' => $_POST['date']
+                );
+                try {
+                    $stmt = $db->prepare($query);
+                    $result = $stmt->execute($query_params);
+                } catch(PDOException $e) {
+                    die("Failed to run query: " . $e->getMessage());
+                }
+                $preBookedTimes = array();
+                while ($alreadyBookedTimes = $stmt->fetch(PDO::FETCH_ASSOC)) {
+                    $times = explode(":", $alreadyBookedTimes["time"]);
+                    array_push($preBookedTimes, intval($times[0]));
+                }
+                // Determine which days go in the select field.
                 for($i = $beginTime; $i < $endTime; $i++) {
+                    $alreadyBooked = false;
                     if ($i < 12) {
-                        echo "<option value =\"" . $i . ":00 am\">" . $i . ":00 am</option>";
-                    } else if ($i == 12) { 
-                        echo "<option value =\"" . $i . ":00 pm\">" . $i . ":00 pm</option>";
+                        foreach($preBookedTimes as $hour) {
+                            if($i == $hour) {
+                                $alreadyBooked = true;
+                                break;
+                            }
+                        }
+                        if (!$alreadyBooked) {
+                            echo "<option value =\"" . $i . ":00 am\">" . $i . ":00 am</option>";
+                        }
+                    } else if ($i == 12) {
+                        foreach($preBookedTimes as $hour) {
+                            if($i == $hour) {
+                                $alreadyBooked = true;
+                                break;
+                            }
+                        }
+                        if(!$alreadyBooked) {
+                            echo "<option value =\"" . $i . ":00 pm\">" . $i . ":00 pm</option>";
+                        }
                     } else if ($i > 12 && $i < 24) {
                         $val = $i - 12;
-                        echo "<option value =\"" . $val . ":00 pm\">" . $val . ":00 pm</option>";   
+                        foreach($preBookedTimes as $hour) {
+                            if($val == $hour) {
+                                $alreadyBooked = true;
+                                break;
+                            }
+                        }
+                        if(!$alreadyBooked) {
+                            echo "<option value =\"" . $val . ":00 pm\">" . $val . ":00 pm</option>";
+                        }
                     } else if ($i == 24) {
                         $val = $i - 12;
-                        echo "<option value =\"" . $val . ":00 am\">" . $val . ":00 am</option>";
+                        foreach($preBookedTimes as $hour) {
+                            if($val == $hour) {
+                                $alreadyBooked = true;
+                                break;
+                            }
+                        }
+                        if(!$alreadyBooked) {
+                            echo "<option value =\"" . $val . ":00 am\">" . $val . ":00 am</option>";
+                        }
                     } else {
                         $val = $i - 24;
-                        echo "<option value =\"" . $val . ":00 am\">" . $val . ":00 am</option>";
+                        foreach($preBookedTimes as $hour) {
+                            if($val == $hour) {
+                                $alreadyBooked = true;
+                                break;
+                            }
+                        }
+                        if(!$alreadyBooked) {
+                            echo "<option value =\"" . $val . ":00 am\">" . $val . ":00 am</option>";
+                        }
                     }
                 }
                 echo "</select><br/><br/>";
